requirements.txt
requests>=2.31.0
PyYAML>=6.0
feedparser>=6.0.10
python-dateutil>=2.8.2